services/interpreTest/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Supabase
supabase>=2.0.0

# NLP & Speech
spacy>=3.7.0
# scispacy>=0.5.0
# negspacy>=1.0.0
openai>=1.3.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Audio processing
librosa>=0.10.0

# Database & Cache
redis>=5.0.0
asyncpg>=0.29.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Utilities
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0

# Jupyter & Development
jupyterlab>=4.0.0
ipykernel>=6.27.0
matplotlib>=3.8.0
pandas>=2.1.0
